chore: Move file from templates to components
